requirements.txt
Pillow==9.0.0
selenium==4.1.0
pynput==1.7.6
PyYAML==6.0
pyalsaaudio==0.9.0
